backend/requirements.txt
Django==4.0.4
django-filter==21.1
django-cors-headers==3.11.0
django-debug-toolbar==3.2.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
Pillow==9.0.1
psycopg2-binary==2.9.3
gunicorn===20.1.0
celery===5.2.6